Timed dataframes

A timed dataframe is a dataframe with a time point (as a time_period vector) associated with every entry. It will also have a date column (as a normal Date object) once converted. This is a line list with a date not a time series.

as.timed_df(
  df,
  days_in_period = NULL,
  start_date = NULL,
  class_col = "class",
  grp_label = "Group {.grp_id}"
)

is.timed_df(df)

Arguments

df

a dataframe containing either a time (as time_period, or numeric) or a date (as a Date) column. This may be grouped in which case the grouping defines individual timeseries, which are generally treated as independent.

days_in_period

the number of days in one time unit. May be non integer. If the time column in the dataframe is already a time_period then this is optional, but if provided will recalibrate the time_period to a new interval. Otherwise it is required

start_date

the zero day. If the time column in the dataframe is already a time_period then this is optional, but if provided will recalibrate the time_period to a new interval. If the dataframe has a date column then this is optional and the zero day is assumed to be the start of the timeseries.

class_col

the column name of a class column if present. Defaults to "class" but will only be used if the column is present. The class column may be a binomial or multinomial class marker.

grp_label

a glue specification to apply to the groups of the df dataframe to create a unique label for the groups.

Value

a dateframe with a time column (as a time_period) and date as a Date. if originally grouped it will have a group column and if class_col was# defined a class column. Other columns are preserved.

Functions

is.timed_df(): Check if this is a timed_df

Examples

library(tidyverse)
#> ── Attaching packages ─────────────────────────────────────── tidyverse 1.3.2 ──
#> ✔ ggplot2 3.3.6      ✔ purrr   0.3.4 
#> ✔ tibble  3.1.8      ✔ dplyr   1.0.10
#> ✔ tidyr   1.2.1      ✔ stringr 1.4.1 
#> ✔ readr   2.1.2      ✔ forcats 0.5.2 
#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::filter() masks stats::filter()
#> ✖ dplyr::lag()    masks stats::lag()
library(growthrates)

# a timed_df object:
tmp1 = tibble(
    time = floor(runif(100, 0, 20)),
    value = rnorm(100)
  ) %>%
  as.timed_df(days_in_period=7,start_date="2020-01-01")
# defaults to `count = n()`

tmp1 %>% time_summarise()
#> # A tibble: 20 × 3
#>    time       count date      
#>  * <time_prd> <int> <date>    
#>  1  0             3 2020-01-01
#>  2  1             5 2020-01-08
#>  3  2             8 2020-01-15
#>  4  3             8 2020-01-22
#>  5  4             7 2020-01-29
#>  6  5             8 2020-02-05
#>  7  6             7 2020-02-12
#>  8  7             6 2020-02-19
#>  9  8             7 2020-02-26
#> 10  9             5 2020-03-04
#> 11 10             6 2020-03-11
#> 12 11             3 2020-03-18
#> 13 12             4 2020-03-25
#> 14 13             3 2020-04-01
#> 15 14             2 2020-04-08
#> 16 15             2 2020-04-15
#> 17 16             2 2020-04-22
#> 18 17             4 2020-04-29
#> 19 18             3 2020-05-06
#> 20 19             7 2020-05-13

tmp1 %>% time_summarise(value = mean(value))
#> # A tibble: 20 × 3
#>    time         value date      
#>  * <time_prd>   <dbl> <date>    
#>  1  0         -0.489  2020-01-01
#>  2  1          0.232  2020-01-08
#>  3  2          0.386  2020-01-15
#>  4  3         -0.0224 2020-01-22
#>  5  4          0.210  2020-01-29
#>  6  5         -0.651  2020-02-05
#>  7  6         -0.108  2020-02-12
#>  8  7         -0.208  2020-02-19
#>  9  8          0.276  2020-02-26
#> 10  9         -0.283  2020-03-04
#> 11 10         -0.241  2020-03-11
#> 12 11         -0.685  2020-03-18
#> 13 12          0.692  2020-03-25
#> 14 13          0.386  2020-04-01
#> 15 14          0.0554 2020-04-08
#> 16 15          1.15   2020-04-15
#> 17 16         -0.831  2020-04-22
#> 18 17         -0.196  2020-04-29
#> 19 18         -0.842  2020-05-06
#> 20 19         -0.387  2020-05-13
# the resulting timeseries are weekly as the inupt was weekly.

# class grouping example

tmp2 = tibble(
    mygroup = c(rep("g1",50),rep("g2",50),rep("g1",50),rep("g2",50)),
    myclass = c(rep("a",100),rep("b",100)),
    time = floor(runif(200, 0, 20)),
    value = rnorm(200)
  ) %>%
  group_by(mygroup) %>%
  as.timed_df(days_in_period=7,start_date="2020-01-01",class_col=myclass,grp_label = "{mygroup}")
#> classes defined: class=myclass
#> groups defined (mygroup): group=`{mygroup}`

tmp3 = tmp2 %>% time_summarise()
tmp3 %>% restore_groups()
#> # A tibble: 80 × 5
#>    mygroup myclass time       count date      
#>    <chr>   <chr>   <time_prd> <int> <date>    
#>  1 g1      a       0              2 2020-01-01
#>  2 g1      a       1              1 2020-01-08
#>  3 g1      a       2              2 2020-01-15
#>  4 g1      a       3              1 2020-01-22
#>  5 g1      a       4              3 2020-01-29
#>  6 g1      a       5              3 2020-02-05
#>  7 g1      a       6              1 2020-02-12
#>  8 g1      a       7              2 2020-02-19
#>  9 g1      a       8              5 2020-02-26
#> 10 g1      a       9              3 2020-03-04
#> # … with 70 more rows

tmp4 = tmp3 %>% time_decompose()
tmp4 %>% restore_groups()
#> # A tibble: 200 × 3
#>    mygroup myclass time      
#>    <chr>   <chr>   <time_prd>
#>  1 g1      a       0         
#>  2 g1      a       0         
#>  3 g1      a       1         
#>  4 g1      a       2         
#>  5 g1      a       2         
#>  6 g1      a       3         
#>  7 g1      a       4         
#>  8 g1      a       4         
#>  9 g1      a       4         
#> 10 g1      a       5         
#> # … with 190 more rows