A time-series is a regular sequence of times with other data

The time series is expected to have at least one of a time or date columns and any other set of data columns. There can be multiple time-series within this if the class column is present or the input is grouped. For every class and time/date combination there can be only one entry. Missing values are not filled in. If there is grouping (but no class column) then it is temporarily stored and replaced with an integer class column. This can be restored with restore_grouping() later

as.timeseries(df, ...)

is.timeseries(df)

is.multinomial_ts(df)

Arguments

df

a dataframe with time or date and optionally class or class defined by grouping

...

Arguments passed on to as.timed_df

days_in_period

the number of days in one time unit. May be non integer. If the time column in the dataframe is already a time_period then this is optional, but if provided will recalibrate the time_period to a new interval. Otherwise it is required

start_date

the zero day. If the time column in the dataframe is already a time_period then this is optional, but if provided will recalibrate the time_period to a new interval. If the dataframe has a date column then this is optional and the zero day is assumed to be the start of the timeseries.

class_col

the column name of a class column if present. Defaults to "class" but will only be used if the column is present. The class column may be a binomial or multinomial class marker.

grp_label

a glue specification to apply to the groups of the df dataframe to create a unique label for the groups.

Value

a timeseries object with groupwise unique time and date columns

Functions

is.timeseries(): Check is a timeseries

is.multinomial_ts(): Check is a mulinomial timeseries

Examples

library(tidyverse)
library(growthrates)

 tibble(time = 0:99, value=rnorm(100)) %>%
   as.timeseries(days_in_period=7,start_date="2020-01-01")
#> # A tibble: 100 × 3
#>    time         value date      
#>  * <time_prd>   <dbl> <date>    
#>  1 0          -2.02   2020-01-01
#>  2 1           0.0394 2020-01-08
#>  3 2           0.0487 2020-01-15
#>  4 3          -1.91   2020-01-22
#>  5 4          -0.848  2020-01-29
#>  6 5          -0.695  2020-02-05
#>  7 6          -1.32   2020-02-12
#>  8 7           0.613  2020-02-19
#>  9 8           0.543  2020-02-26
#> 10 9           0.780  2020-03-04
#> # … with 90 more rows

 # defining a multinomial with a class column
 tmp1 = tibble(
     class = c(rep(1,100),rep(2,100)),
     time = rep(0:99,2),
     value = rnorm(200)
  ) %>% as.timeseries(days_in_period=7,start_date="2020-01-01")
#> classes defined: class=class

 restore_groups(tmp1)
#> # A tibble: 200 × 4
#>    class time          value date      
#>  * <dbl> <time_prd>    <dbl> <date>    
#>  1     1 0           1.36    2020-01-01
#>  2     1 1          -0.00439 2020-01-08
#>  3     1 2          -0.0841  2020-01-15
#>  4     1 3           0.572   2020-01-22
#>  5     1 4          -1.14    2020-01-29
#>  6     1 5           0.900   2020-02-05
#>  7     1 6           0.220   2020-02-12
#>  8     1 7           0.351   2020-02-19
#>  9     1 8          -0.587   2020-02-26
#> 10     1 9           1.04    2020-03-04
#> # … with 190 more rows
 # still a timeseries:
 class(restore_groups(tmp1))
#> [1] "timeseries" "timed_df"   "tbl_df"     "tbl"        "data.frame"

 # defining a multinomial with a grouing
 tmp = tibble(
     mygroup = c(rep("a",100),rep("b",100)),
     time = rep(0:99,2),
     value = rnorm(200)
   ) %>%
   group_by(mygroup) %>%
   as.timeseries(days_in_period=7,start_date="2020-01-01")
#> groups defined (mygroup): group=`Group {.grp_id}`

 restore_groups(tmp)
#> # A tibble: 200 × 4
#>    mygroup time          value date      
#>    <chr>   <time_prd>    <dbl> <date>    
#>  1 a       0           1.96    2020-01-01
#>  2 a       1           0.510   2020-01-08
#>  3 a       2           2.41    2020-01-15
#>  4 a       3          -0.276   2020-01-22
#>  5 a       4          -1.40    2020-01-29
#>  6 a       5          -2.18    2020-02-05
#>  7 a       6          -0.257   2020-02-12
#>  8 a       7           0.00479 2020-02-19
#>  9 a       8           2.04    2020-02-26
#> 10 a       9           0.539   2020-03-04
#> # … with 190 more rows
 # no longer a time series:
 class(restore_groups(tmp))
#> [1] "tbl_df"     "tbl"        "data.frame"