Places a set of dates within a regular time series

The counterpart to full_seq_dates(). Take an original set of data and place it within a regular time series where the periodicity of the time series may be expressed as numbers of days, weeks, months quarters, or years, and the periods are defined by an anchoring date, day of the week or by reference to the start or end of the input dates. This can either return the periods as dates or factors (e.g. for plotting) or as a time_period for analysis that relies on a numeric representation of the date or duration from the anchor.

cut_date(
  dates,
  unit,
  anchor = "start",
  output = c("date", "factor", "time_period"),
  dfmt = "%d/%b/%y",
  ifmt = "%s — %s",
  ...
)

Arguments

dates

a set of dates

unit

a period e.g. "1 week"

anchor

one of a date, "start" or "end" or a weekday name e.g. "mon" this will always be one of the start of the time periods we are cutting into

output

return the result as either a "date" (the default), an ordered "factor" with the date ranges as a label, or as a "time_period". The result is named with labels referring to the

dfmt

the strptime format for the dates in the labels

ifmt

a sprintf format for the period label containing %s exactly twice.

...

ignored

Value

a set of dates, times or a factor level, representing the start of the period the date falls into, where the period is defined by the duration and the anchor

Examples

dates = as.Date(c("2020-01-01","2020-02-01","2020-01-15","2020-02-03",NA))
fs = tidyr::full_seq(dates, "2 days")
dates - cut_date(dates, "2 days")
#> Time differences in days
#> 01/Jan/20 — 02/Jan/20 31/Jan/20 — 01/Feb/20 15/Jan/20 — 16/Jan/20 
#>                     0                     1                     0 
#> 02/Feb/20 — 03/Feb/20               Unknown 
#>                     1                    NA 
cut_date(dates,unit="2 days", output="time_period")
#> time unit: 2d 0H 0M 0S, origin: 2020-01-01 (a Wednesday)
#> [1]  0 15  7 16 NA

# A weekly set of dates:
dates2 = Sys.Date() + floor(stats::runif(50,max=10))*7

# in this specific situation the final date is not truncated because the
# input data is seen as an exact match for the whole output period.
cut_date(dates2, "1 week", "sun", output="factor")
#>  [1] 16/Apr/23 — 22/Apr/23 16/Apr/23 — 22/Apr/23 19/Mar/23 — 25/Mar/23
#>  [4] 19/Mar/23 — 25/Mar/23 23/Apr/23 — 29/Apr/23 02/Apr/23 — 08/Apr/23
#>  [7] 02/Apr/23 — 08/Apr/23 26/Feb/23 — 04/Mar/23 19/Mar/23 — 25/Mar/23
#> [10] 09/Apr/23 — 15/Apr/23 26/Mar/23 — 01/Apr/23 19/Feb/23 — 25/Feb/23
#> [13] 19/Feb/23 — 25/Feb/23 12/Mar/23 — 18/Mar/23 19/Mar/23 — 25/Mar/23
#> [16] 09/Apr/23 — 15/Apr/23 26/Mar/23 — 01/Apr/23 09/Apr/23 — 15/Apr/23
#> [19] 12/Mar/23 — 18/Mar/23 19/Mar/23 — 25/Mar/23 19/Mar/23 — 25/Mar/23
#> [22] 05/Mar/23 — 11/Mar/23 26/Mar/23 — 01/Apr/23 12/Mar/23 — 18/Mar/23
#> [25] 23/Apr/23 — 29/Apr/23 26/Mar/23 — 01/Apr/23 19/Feb/23 — 25/Feb/23
#> [28] 12/Mar/23 — 18/Mar/23 26/Feb/23 — 04/Mar/23 26/Mar/23 — 01/Apr/23
#> [31] 19/Mar/23 — 25/Mar/23 16/Apr/23 — 22/Apr/23 09/Apr/23 — 15/Apr/23
#> [34] 09/Apr/23 — 15/Apr/23 12/Mar/23 — 18/Mar/23 19/Mar/23 — 25/Mar/23
#> [37] 23/Apr/23 — 29/Apr/23 02/Apr/23 — 08/Apr/23 19/Feb/23 — 25/Feb/23
#> [40] 09/Apr/23 — 15/Apr/23 26/Mar/23 — 01/Apr/23 09/Apr/23 — 15/Apr/23
#> [43] 12/Mar/23 — 18/Mar/23 12/Mar/23 — 18/Mar/23 23/Apr/23 — 29/Apr/23
#> [46] 05/Mar/23 — 11/Mar/23 26/Mar/23 — 01/Apr/23 16/Apr/23 — 22/Apr/23
#> [49] 05/Mar/23 — 11/Mar/23 02/Apr/23 — 08/Apr/23
#> 11 Levels: 19/Feb/23 — 25/Feb/23 < ... < 30/Apr/23 — 06/May/23
cut_date(dates2, dfmt = "%d/%b", output="factor", unit = "2 weeks", anchor="sun")
#>  [1] 16/Apr — 29/Apr 16/Apr — 29/Apr 19/Mar — 01/Apr 19/Mar — 01/Apr
#>  [5] 16/Apr — 29/Apr 02/Apr — 15/Apr 02/Apr — 15/Apr 19/Feb — 04/Mar
#>  [9] 19/Mar — 01/Apr 02/Apr — 15/Apr 19/Mar — 01/Apr 19/Feb — 04/Mar
#> [13] 19/Feb — 04/Mar 05/Mar — 18/Mar 19/Mar — 01/Apr 02/Apr — 15/Apr
#> [17] 19/Mar — 01/Apr 02/Apr — 15/Apr 05/Mar — 18/Mar 19/Mar — 01/Apr
#> [21] 19/Mar — 01/Apr 05/Mar — 18/Mar 19/Mar — 01/Apr 05/Mar — 18/Mar
#> [25] 16/Apr — 29/Apr 19/Mar — 01/Apr 19/Feb — 04/Mar 05/Mar — 18/Mar
#> [29] 19/Feb — 04/Mar 19/Mar — 01/Apr 19/Mar — 01/Apr 16/Apr — 29/Apr
#> [33] 02/Apr — 15/Apr 02/Apr — 15/Apr 05/Mar — 18/Mar 19/Mar — 01/Apr
#> [37] 16/Apr — 29/Apr 02/Apr — 15/Apr 19/Feb — 04/Mar 02/Apr — 15/Apr
#> [41] 19/Mar — 01/Apr 02/Apr — 15/Apr 05/Mar — 18/Mar 05/Mar — 18/Mar
#> [45] 16/Apr — 29/Apr 05/Mar — 18/Mar 19/Mar — 01/Apr 16/Apr — 29/Apr
#> [49] 05/Mar — 18/Mar 02/Apr — 15/Apr
#> 6 Levels: 19/Feb — 04/Mar < 05/Mar — 18/Mar < ... < 30/Apr — 13/May