Estimate the probabilities of a multinomial class based on count data

takes a list of times, classes and counts, e.g. a COGUK variant like data set with time, (multinomial) class (e.g. variant) and count being the count in that time period. Fits a quadratic B-spline on time to the proportion of the data using nnet::multinom.

estimate_multinomial_proportion(timeseries, deg = 2, ...)

Arguments

timeseries

a timeseries of class and count data where class has 2 or more levels. The classwise counts are used to assess a time varying probability using spline terms.

deg

the degree of the polynomial spline

...

Arguments passed on to splines::bs

x

the predictor variable. Missing values are allowed.

df

degrees of freedom; one can specify df rather than knots; bs() then chooses df-degree (minus one if there is an intercept) knots at suitable quantiles of x (which will ignore missing values). The default, NULL, takes the number of inner knots as length(knots). If that is zero as per default, that corresponds to df = degree - intercept.

knots

the internal breakpoints that define the spline. The default is NULL, which results in a basis for ordinary polynomial regression. Typical values are the mean or median for one knot, quantiles for more knots. See also Boundary.knots.

degree

degree of the piecewise polynomial---default is 3 for cubic splines.

intercept

if TRUE, an intercept is included in the basis; default is FALSE.

Boundary.knots

boundary points at which to anchor the B-spline basis (default the range of the non-NA data). If both knots and Boundary.knots are supplied, the basis parameters do not depend on x. Data can extend beyond Boundary.knots.

Value

a timeseries of class and proportion columns where the proportion is an estimate of the multinomial class probability at any given time.

Examples

if (FALSE) {
  utils::vignette("variant-proportions", package="growthrates")
}