Label a time period

Create a set of labels for a time period based on the start and duration of the period. The format is configurable using the start and end dates and the dfmt and ifmt parameters, however if the time period has names then these are used in preference.

# S3 method for time_period
labels(
  object,
  ...,
  dfmt = "%d/%b",
  ifmt = "{start} — {end}",
  na.value = "Unknown"
)

Arguments

object

a set of decimal times as a time_period

...

not used

dfmt

a strptime format specification for the format of the date

ifmt

a glue spec referring to start and end of the period as a formatted date

na.value

a label for NA times

Value

a set of character labels for the time

Examples

eg = as.time_period(Sys.Date()+0:10*7, anchor="start")
#> No unit given. Guessing a sensible value from the dates gives: 7d 0H 0M 0S
labels(eg)
#> 26/Dec — 01/Jan
#> 02/Jan — 08/Jan
#> 09/Jan — 15/Jan
#> 16/Jan — 22/Jan
#> 23/Jan — 29/Jan
#> 30/Jan — 05/Feb
#> 06/Feb — 12/Feb
#> 13/Feb — 19/Feb
#> 20/Feb — 26/Feb
#> 27/Feb — 04/Mar
#> 05/Mar — 11/Mar
labels(eg, ifmt="{start}", dfmt="%d/%b/%y")
#> 26/Dec/23
#> 02/Jan/24
#> 09/Jan/24
#> 16/Jan/24
#> 23/Jan/24
#> 30/Jan/24
#> 06/Feb/24
#> 13/Feb/24
#> 20/Feb/24
#> 27/Feb/24
#> 05/Mar/24
labels(eg, ifmt="until {end}", dfmt="%d %b %Y")
#> until 01 Jan 2024
#> until 08 Jan 2024
#> until 15 Jan 2024
#> until 22 Jan 2024
#> until 29 Jan 2024
#> until 05 Feb 2024
#> until 12 Feb 2024
#> until 19 Feb 2024
#> until 26 Feb 2024
#> until 04 Mar 2024
#> until 11 Mar 2024

# labels retained in constructor:
eg2 = Sys.Date()+0:10*7
names(eg2) = paste0("week ",0:10)
labels(eg2)
#>  [1] "week 0"  "week 1"  "week 2"  "week 3"  "week 4"  "week 5"  "week 6" 
#>  [8] "week 7"  "week 8"  "week 9"  "week 10"
labels(as.time_period(eg2, anchor="start"))
#> No unit given. Guessing a sensible value from the dates gives: 7d 0H 0M 0S
#>  [1] "week 0"  "week 1"  "week 2"  "week 3"  "week 4"  "week 5"  "week 6" 
#>  [8] "week 7"  "week 8"  "week 9"  "week 10"