A binomial proportion estimate and associated exponential growth rate

takes a list of times and counts based on and fits a quasi-binomial model fitted with a logit link function to proportion data using local regression.

proportion_locfit_model(
  d,
  ...,
  window = 14,
  deg = 2,
  frequency = "1 day",
  predict = TRUE
)

Arguments

d

input dataframe with: time (as a time_period), count, and denom columns. This must be ungrouped (but this function can be used directly within a group_modify)

...

not used and present to allow proportion model to be used in a group_modify

window

a number of data points defining the bandwidth of the estimate, smaller values result in less smoothing, large value in more.

deg

polynomial degree

frequency

the desired time unit between the output estimates.

predict

result a prediction. If false we return the model.

Value

a new dataframe with columns time (as a time period), proportion.0.5

plus quantiles, and relative.growth.0.5 (plus quantiles) columns or a list of 2 locfit model objects (one for proportion and one for relative growth rate)

Details

This expects d to contain one combination of:

time and count and denom columns - e.g. all tests conducted.

This results is a one versus others comparison binomial proportion estimate plus a relative growth rate estimate.

Examples

# Proportions
# data4 %>%
#   mutate(time = cut_date(date, unit = "1 week", anchor="start", output="time_period")) %>%
#   group_by(time, serotype) %>%
#   summarise(count = n(), .groups="drop_last") %>%
#   mutate(denom = sum(count)) %>%
#   group_by(serotype) %>%
#   group_modify(proportion_model, deg=3) %>%
#   glimpse()

# Counts
# data4 %>%
#   mutate(time = cut_date(date, unit = "1 week", anchor="start", output="time_period")) %>%
#   group_by(time, serotype) %>%
#   summarise(count = n(), .groups="drop_last") %>%
#   group_by(serotype) %>%
#   group_modify(poisson_model, df=4) %>%
#   glimpse()