Summarise data from a timed_df to a timeseries

Summarise data from a timed_df to a timeseries

time_summarise(timed_df, ..., .fill = list())

Arguments

timed_df

a line list of data you want to summarise by the time column. If this is grouped the grouping is preserved and converted into a class column in the output. Ff the timed_df has a class column that is used in preference.

...

a spec for a dplyr::summary(...) - optional, and if not provided a count = dplyr::n() or a count = sum(count) is performed.

.fill

a list similar to tidyr::complete for values to fill variables with

Value

a timeseries dataframe, by time

Examples

library(tidyverse)
library(growthrates)

# a timed_df object:
tmp1 = tibble(
    time = floor(runif(100, 0, 20)),
    value = rnorm(100)
  ) %>%
  as.timed_df(days_in_period=7,start_date="2020-01-01")
# defaults to `count = n()`

tmp1 %>% time_summarise()
#> # A tibble: 20 × 3
#>    time       count date      
#>  * <time_prd> <int> <date>    
#>  1  0             3 2020-01-01
#>  2  1             3 2020-01-08
#>  3  2             3 2020-01-15
#>  4  3             3 2020-01-22
#>  5  4             6 2020-01-29
#>  6  5             3 2020-02-05
#>  7  6             5 2020-02-12
#>  8  7             7 2020-02-19
#>  9  8             6 2020-02-26
#> 10  9             3 2020-03-04
#> 11 10            11 2020-03-11
#> 12 11             6 2020-03-18
#> 13 12             5 2020-03-25
#> 14 13             6 2020-04-01
#> 15 14             7 2020-04-08
#> 16 15             5 2020-04-15
#> 17 16             6 2020-04-22
#> 18 17             1 2020-04-29
#> 19 18             5 2020-05-06
#> 20 19             6 2020-05-13

tmp1 %>% time_summarise(value = mean(value))
#> # A tibble: 20 × 3
#>    time         value date      
#>  * <time_prd>   <dbl> <date>    
#>  1  0         -0.286  2020-01-01
#>  2  1          0.823  2020-01-08
#>  3  2         -1.38   2020-01-15
#>  4  3         -0.190  2020-01-22
#>  5  4         -0.0477 2020-01-29
#>  6  5         -0.629  2020-02-05
#>  7  6          0.0105 2020-02-12
#>  8  7          0.442  2020-02-19
#>  9  8          0.250  2020-02-26
#> 10  9          0.306  2020-03-04
#> 11 10         -0.0997 2020-03-11
#> 12 11          0.297  2020-03-18
#> 13 12         -0.649  2020-03-25
#> 14 13         -0.161  2020-04-01
#> 15 14          0.195  2020-04-08
#> 16 15          0.240  2020-04-15
#> 17 16         -0.386  2020-04-22
#> 18 17         -0.308  2020-04-29
#> 19 18          0.126  2020-05-06
#> 20 19          0.181  2020-05-13
# the resulting timeseries are weekly as the inupt was weekly.

# class grouping example

tmp2 = tibble(
    mygroup = c(rep("g1",50),rep("g2",50),rep("g1",50),rep("g2",50)),
    myclass = c(rep("a",100),rep("b",100)),
    time = floor(runif(200, 0, 20)),
    value = rnorm(200)
  ) %>%
  group_by(mygroup) %>%
  as.timed_df(days_in_period=7,start_date="2020-01-01",class_col=myclass,grp_label = "{mygroup}")
#> classes defined: class=myclass
#> groups defined (mygroup): group=`{mygroup}`

tmp3 = tmp2 %>% time_summarise()
tmp3 %>% restore_groups()
#> # A tibble: 80 × 5
#>    mygroup myclass time       count date      
#>    <chr>   <chr>   <time_prd> <int> <date>    
#>  1 g1      a       0              5 2020-01-01
#>  2 g1      a       1              1 2020-01-08
#>  3 g1      a       2              3 2020-01-15
#>  4 g1      a       3              1 2020-01-22
#>  5 g1      a       4              2 2020-01-29
#>  6 g1      a       5              1 2020-02-05
#>  7 g1      a       6              3 2020-02-12
#>  8 g1      a       7              6 2020-02-19
#>  9 g1      a       8              1 2020-02-26
#> 10 g1      a       9              2 2020-03-04
#> # … with 70 more rows

tmp4 = tmp3 %>% time_decompose()
tmp4 %>% restore_groups()
#> # A tibble: 200 × 3
#>    mygroup myclass time      
#>    <chr>   <chr>   <time_prd>
#>  1 g1      a       0         
#>  2 g1      a       0         
#>  3 g1      a       0         
#>  4 g1      a       0         
#>  5 g1      a       0         
#>  6 g1      a       1         
#>  7 g1      a       2         
#>  8 g1      a       2         
#>  9 g1      a       2         
#> 10 g1      a       3         
#> # … with 190 more rows