Places a set of dates within a regular time series

The counterpart to date_seq_dates(). Take an original set of data and place it within a regular time series where the periodicity of the time series may be expressed as numbers of days, weeks, months quarters, or years, and the periods are defined by an anchoring date, day of the week or by reference to the start or end of the input dates. This can either return the periods as dates or factors (e.g. for plotting) or as a time_period for analysis that relies on a numeric representation of the date or duration from the anchor.

Usage

cut_date(
  dates,
  unit,
  anchor = "start",
  output = c("date", "factor", "time_period"),
  dfmt = "%d/%b/%y",
  ifmt = "{start} — {end}",
  ...
)

Arguments

dates

a set of dates

unit

a period e.g. "1 week"

anchor

one of a date, "start" or "end" or a weekday name e.g. "mon" this will always be one of the start of the time periods we are cutting into

output

return the result as either a "date" (the default), an ordered "factor" with the date ranges as a label, or as a "time_period". The result is named with labels referring to the

dfmt

the strptime format for the dates in the labels

ifmt

a sprintf format for the period label containing %s exactly twice.

...

ignored

Value

a set of dates, times or a factor level, representing the start of the period the date falls into, where the period is defined by the duration and the anchor

Examples

dates = as.Date(c("2020-01-01","2020-02-01","2020-01-15","2020-02-03",NA))
fs = growthrates::date_seq(dates, "2 days")
dates - cut_date(dates, "2 days")
#> Time differences in days
#> 01/Jan/20 — 02/Jan/20 31/Jan/20 — 01/Feb/20 15/Jan/20 — 16/Jan/20 
#>                     0                     1                     0 
#> 02/Feb/20 — 03/Feb/20               Unknown 
#>                     1                    NA 
cut_date(dates,unit="2 days", output="time_period")
#> time unit: 2 days, origin: 2020-01-01 (a Wednesday)
#> [1]  0 15  7 16 NA

# A weekly set of dates:
dates2 = Sys.Date() + floor(stats::runif(50,max=10))*7

# in this specific situation the final date is not truncated because the
# input data is seen as an exact match for the whole output period.
cut_date(dates2, "1 week", "sun", output="factor")
#>  [1] 21/Apr/24 — 27/Apr/24 16/Jun/24 — 22/Jun/24 02/Jun/24 — 08/Jun/24
#>  [4] 28/Apr/24 — 04/May/24 21/Apr/24 — 27/Apr/24 19/May/24 — 25/May/24
#>  [7] 19/May/24 — 25/May/24 05/May/24 — 11/May/24 09/Jun/24 — 15/Jun/24
#> [10] 09/Jun/24 — 15/Jun/24 16/Jun/24 — 22/Jun/24 28/Apr/24 — 04/May/24
#> [13] 21/Apr/24 — 27/Apr/24 12/May/24 — 18/May/24 19/May/24 — 25/May/24
#> [16] 28/Apr/24 — 04/May/24 19/May/24 — 25/May/24 21/Apr/24 — 27/Apr/24
#> [19] 12/May/24 — 18/May/24 23/Jun/24 — 29/Jun/24 05/May/24 — 11/May/24
#> [22] 02/Jun/24 — 08/Jun/24 09/Jun/24 — 15/Jun/24 28/Apr/24 — 04/May/24
#> [25] 23/Jun/24 — 29/Jun/24 09/Jun/24 — 15/Jun/24 21/Apr/24 — 27/Apr/24
#> [28] 26/May/24 — 01/Jun/24 02/Jun/24 — 08/Jun/24 02/Jun/24 — 08/Jun/24
#> [31] 21/Apr/24 — 27/Apr/24 05/May/24 — 11/May/24 12/May/24 — 18/May/24
#> [34] 02/Jun/24 — 08/Jun/24 19/May/24 — 25/May/24 19/May/24 — 25/May/24
#> [37] 09/Jun/24 — 15/Jun/24 23/Jun/24 — 29/Jun/24 28/Apr/24 — 04/May/24
#> [40] 05/May/24 — 11/May/24 02/Jun/24 — 08/Jun/24 19/May/24 — 25/May/24
#> [43] 02/Jun/24 — 08/Jun/24 02/Jun/24 — 08/Jun/24 21/Apr/24 — 27/Apr/24
#> [46] 09/Jun/24 — 15/Jun/24 09/Jun/24 — 15/Jun/24 23/Jun/24 — 29/Jun/24
#> [49] 23/Jun/24 — 29/Jun/24 12/May/24 — 18/May/24
#> 11 Levels: 21/Apr/24 — 27/Apr/24 < ... < 30/Jun/24 — 06/Jul/24
cut_date(dates2, dfmt = "%d/%b", output="factor", unit = "2 weeks", anchor="sun")
#>  [1] 21/Apr — 04/May 16/Jun — 29/Jun 02/Jun — 15/Jun 21/Apr — 04/May
#>  [5] 21/Apr — 04/May 19/May — 01/Jun 19/May — 01/Jun 05/May — 18/May
#>  [9] 02/Jun — 15/Jun 02/Jun — 15/Jun 16/Jun — 29/Jun 21/Apr — 04/May
#> [13] 21/Apr — 04/May 05/May — 18/May 19/May — 01/Jun 21/Apr — 04/May
#> [17] 19/May — 01/Jun 21/Apr — 04/May 05/May — 18/May 16/Jun — 29/Jun
#> [21] 05/May — 18/May 02/Jun — 15/Jun 02/Jun — 15/Jun 21/Apr — 04/May
#> [25] 16/Jun — 29/Jun 02/Jun — 15/Jun 21/Apr — 04/May 19/May — 01/Jun
#> [29] 02/Jun — 15/Jun 02/Jun — 15/Jun 21/Apr — 04/May 05/May — 18/May
#> [33] 05/May — 18/May 02/Jun — 15/Jun 19/May — 01/Jun 19/May — 01/Jun
#> [37] 02/Jun — 15/Jun 16/Jun — 29/Jun 21/Apr — 04/May 05/May — 18/May
#> [41] 02/Jun — 15/Jun 19/May — 01/Jun 02/Jun — 15/Jun 02/Jun — 15/Jun
#> [45] 21/Apr — 04/May 02/Jun — 15/Jun 02/Jun — 15/Jun 16/Jun — 29/Jun
#> [49] 16/Jun — 29/Jun 05/May — 18/May
#> 6 Levels: 21/Apr — 04/May < 05/May — 18/May < ... < 30/Jun — 13/Jul