Places a set of dates within a regular time series

The counterpart to full_seq_dates(). Take an original set of data and place it within a regular time series where the periodicity of the time series may be expressed as numbers of days, weeks, months quarters, or years, and the periods are defined by an anchoring date, day of the week or by reference to the start or end of the input dates. This can either return the periods as dates or factors (e.g. for plotting) or as a time_period for analysis that relies on a numeric representation of the date or duration from the anchor.

cut_date(
  dates,
  unit,
  anchor = "start",
  output = c("date", "factor", "time_period"),
  dfmt = "%d/%b/%y",
  ifmt = "%s — %s",
  ...
)

Arguments

dates

a set of dates

unit

a period e.g. "1 week"

anchor

one of a date, "start" or "end" or a weekday name e.g. "mon" this will always be one of the start of the time periods we are cutting into

output

return the result as either a "date" (the default), an ordered "factor" with the date ranges as a label, or as a "time_period". The result is named with labels referring to the

dfmt

the strptime format for the dates in the labels

ifmt

a sprintf format for the period label containing %s exactly twice.

...

ignored

Value

a set of dates, times or a factor level, representing the start of the period the date falls into, where the period is defined by the duration and the anchor

Examples

dates = as.Date(c("2020-01-01","2020-02-01","2020-01-15","2020-02-03",NA))
fs = full_seq(dates, "2 days")
#> Error in full_seq(as.numeric(x), period, tol): `period` must be a number, not the string "2 days".
dates - cut_date(dates, "2 days")
#> Time differences in days
#> 01/Jan/20 — 02/Jan/20 31/Jan/20 — 01/Feb/20 15/Jan/20 — 16/Jan/20 
#>                     0                     1                     0 
#> 02/Feb/20 — 03/Feb/20               Unknown 
#>                     1                    NA 
cut_date(dates,unit="2 days", output="time_period")
#> time unit: 2d 0H 0M 0S, origin: 2020-01-01 (a Wednesday)
#> [1]  0 15  7 16 NA

# A weekly set of dates:
dates2 = Sys.Date() + floor(stats::runif(50,max=10))*7

# in this specific situation the final date is not truncated because the
# input data is seen as an exact match for the whole output period.
cut_date(dates2, "1 week", "sun", output="factor")
#>  [1] 29/Oct/23 — 04/Nov/23 26/Nov/23 — 02/Dec/23 26/Nov/23 — 02/Dec/23
#>  [4] 22/Oct/23 — 28/Oct/23 19/Nov/23 — 25/Nov/23 03/Dec/23 — 09/Dec/23
#>  [7] 08/Oct/23 — 14/Oct/23 19/Nov/23 — 25/Nov/23 03/Dec/23 — 09/Dec/23
#> [10] 03/Dec/23 — 09/Dec/23 15/Oct/23 — 21/Oct/23 19/Nov/23 — 25/Nov/23
#> [13] 05/Nov/23 — 11/Nov/23 26/Nov/23 — 02/Dec/23 12/Nov/23 — 18/Nov/23
#> [16] 19/Nov/23 — 25/Nov/23 15/Oct/23 — 21/Oct/23 26/Nov/23 — 02/Dec/23
#> [19] 15/Oct/23 — 21/Oct/23 22/Oct/23 — 28/Oct/23 05/Nov/23 — 11/Nov/23
#> [22] 08/Oct/23 — 14/Oct/23 08/Oct/23 — 14/Oct/23 03/Dec/23 — 09/Dec/23
#> [25] 26/Nov/23 — 02/Dec/23 05/Nov/23 — 11/Nov/23 12/Nov/23 — 18/Nov/23
#> [28] 05/Nov/23 — 11/Nov/23 12/Nov/23 — 18/Nov/23 03/Dec/23 — 09/Dec/23
#> [31] 03/Dec/23 — 09/Dec/23 29/Oct/23 — 04/Nov/23 08/Oct/23 — 14/Oct/23
#> [34] 03/Dec/23 — 09/Dec/23 15/Oct/23 — 21/Oct/23 01/Oct/23 — 07/Oct/23
#> [37] 01/Oct/23 — 07/Oct/23 03/Dec/23 — 09/Dec/23 03/Dec/23 — 09/Dec/23
#> [40] 01/Oct/23 — 07/Oct/23 12/Nov/23 — 18/Nov/23 01/Oct/23 — 07/Oct/23
#> [43] 15/Oct/23 — 21/Oct/23 19/Nov/23 — 25/Nov/23 03/Dec/23 — 09/Dec/23
#> [46] 29/Oct/23 — 04/Nov/23 15/Oct/23 — 21/Oct/23 19/Nov/23 — 25/Nov/23
#> [49] 05/Nov/23 — 11/Nov/23 01/Oct/23 — 07/Oct/23
#> 10 Levels: 01/Oct/23 — 07/Oct/23 < ... < 03/Dec/23 — 09/Dec/23
cut_date(dates2, dfmt = "%d/%b", output="factor", unit = "2 weeks", anchor="sun")
#>  [1] 29/Oct — 11/Nov 26/Nov — 09/Dec 26/Nov — 09/Dec 15/Oct — 28/Oct
#>  [5] 12/Nov — 25/Nov 26/Nov — 09/Dec 01/Oct — 14/Oct 12/Nov — 25/Nov
#>  [9] 26/Nov — 09/Dec 26/Nov — 09/Dec 15/Oct — 28/Oct 12/Nov — 25/Nov
#> [13] 29/Oct — 11/Nov 26/Nov — 09/Dec 12/Nov — 25/Nov 12/Nov — 25/Nov
#> [17] 15/Oct — 28/Oct 26/Nov — 09/Dec 15/Oct — 28/Oct 15/Oct — 28/Oct
#> [21] 29/Oct — 11/Nov 01/Oct — 14/Oct 01/Oct — 14/Oct 26/Nov — 09/Dec
#> [25] 26/Nov — 09/Dec 29/Oct — 11/Nov 12/Nov — 25/Nov 29/Oct — 11/Nov
#> [29] 12/Nov — 25/Nov 26/Nov — 09/Dec 26/Nov — 09/Dec 29/Oct — 11/Nov
#> [33] 01/Oct — 14/Oct 26/Nov — 09/Dec 15/Oct — 28/Oct 01/Oct — 14/Oct
#> [37] 01/Oct — 14/Oct 26/Nov — 09/Dec 26/Nov — 09/Dec 01/Oct — 14/Oct
#> [41] 12/Nov — 25/Nov 01/Oct — 14/Oct 15/Oct — 28/Oct 12/Nov — 25/Nov
#> [45] 26/Nov — 09/Dec 29/Oct — 11/Nov 15/Oct — 28/Oct 12/Nov — 25/Nov
#> [49] 29/Oct — 11/Nov 01/Oct — 14/Oct
#> 6 Levels: 01/Oct — 14/Oct < 15/Oct — 28/Oct < ... < 10/Dec — 23/Dec