Label a time period

Create a set of labels for a time period based on the start and duration of the period. The format is configurable using the start and end dates and the dfmt and ifmt parameters, however if the time period has names then these are used in preference.

Usage

# S3 method for time_period
labels(
  object,
  ...,
  dfmt = "%d/%b",
  ifmt = "{start} — {end}",
  na.value = "Unknown"
)

Arguments

object

a set of decimal times as a time_period

...

not used

dfmt

a strptime format specification for the format of the date

ifmt

a glue spec referring to start and end of the period as a formatted date

na.value

a label for NA times

Value

a set of character labels for the time

Examples

eg = as.time_period(Sys.Date()+0:10*7, anchor="start")
#> No unit given. Guessing a sensible value from the dates gives: 7d 0H 0M 0S
labels(eg)
#> 05/Jan — 11/Jan
#> 12/Jan — 18/Jan
#> 19/Jan — 25/Jan
#> 26/Jan — 01/Feb
#> 02/Feb — 08/Feb
#> 09/Feb — 15/Feb
#> 16/Feb — 22/Feb
#> 23/Feb — 29/Feb
#> 01/Mar — 07/Mar
#> 08/Mar — 14/Mar
#> 15/Mar — 21/Mar
labels(eg, ifmt="{start}", dfmt="%d/%b/%y")
#> 05/Jan/24
#> 12/Jan/24
#> 19/Jan/24
#> 26/Jan/24
#> 02/Feb/24
#> 09/Feb/24
#> 16/Feb/24
#> 23/Feb/24
#> 01/Mar/24
#> 08/Mar/24
#> 15/Mar/24
labels(eg, ifmt="until {end}", dfmt="%d %b %Y")
#> until 11 Jan 2024
#> until 18 Jan 2024
#> until 25 Jan 2024
#> until 01 Feb 2024
#> until 08 Feb 2024
#> until 15 Feb 2024
#> until 22 Feb 2024
#> until 29 Feb 2024
#> until 07 Mar 2024
#> until 14 Mar 2024
#> until 21 Mar 2024

# labels retained in constructor:
eg2 = Sys.Date()+0:10*7
names(eg2) = paste0("week ",0:10)
labels(eg2)
#>  [1] "week 0"  "week 1"  "week 2"  "week 3"  "week 4"  "week 5"  "week 6" 
#>  [8] "week 7"  "week 8"  "week 9"  "week 10"
labels(as.time_period(eg2, anchor="start"))
#> No unit given. Guessing a sensible value from the dates gives: 7d 0H 0M 0S
#>  [1] "week 0"  "week 1"  "week 2"  "week 3"  "week 4"  "week 5"  "week 6" 
#>  [8] "week 7"  "week 8"  "week 9"  "week 10"