Label a time period

Create a set of labels for a time period based on the start and duration of the period. The format is configurable using the start and end dates and the dfmt and ifmt parameters, however if the time period has names then these are used in preference.

Usage

# S3 method for time_period
labels(
  object,
  ...,
  dfmt = "%d/%b",
  ifmt = "{start} — {end}",
  na.value = "Unknown"
)

Arguments

object

a set of decimal times as a time_period

...

not used

dfmt

a strptime format specification for the format of the date

ifmt

a glue spec referring to start and end of the period as a formatted date

na.value

a label for NA times

Value

a set of character labels for the time

Examples

eg = as.time_period(Sys.Date()+0:10*7, anchor="start")
#> No unit given. Guessing a sensible value from the dates gives: 7d 0H 0M 0S
labels(eg)
#> 26/Apr — 02/May
#> 03/May — 09/May
#> 10/May — 16/May
#> 17/May — 23/May
#> 24/May — 30/May
#> 31/May — 06/Jun
#> 07/Jun — 13/Jun
#> 14/Jun — 20/Jun
#> 21/Jun — 27/Jun
#> 28/Jun — 04/Jul
#> 05/Jul — 11/Jul
labels(eg, ifmt="{start}", dfmt="%d/%b/%y")
#> 26/Apr/24
#> 03/May/24
#> 10/May/24
#> 17/May/24
#> 24/May/24
#> 31/May/24
#> 07/Jun/24
#> 14/Jun/24
#> 21/Jun/24
#> 28/Jun/24
#> 05/Jul/24
labels(eg, ifmt="until {end}", dfmt="%d %b %Y")
#> until 02 May 2024
#> until 09 May 2024
#> until 16 May 2024
#> until 23 May 2024
#> until 30 May 2024
#> until 06 Jun 2024
#> until 13 Jun 2024
#> until 20 Jun 2024
#> until 27 Jun 2024
#> until 04 Jul 2024
#> until 11 Jul 2024

# labels retained in constructor:
eg2 = Sys.Date()+0:10*7
names(eg2) = paste0("week ",0:10)
labels(eg2)
#>  [1] "week 0"  "week 1"  "week 2"  "week 3"  "week 4"  "week 5"  "week 6" 
#>  [8] "week 7"  "week 8"  "week 9"  "week 10"
labels(as.time_period(eg2, anchor="start"))
#> No unit given. Guessing a sensible value from the dates gives: 7d 0H 0M 0S
#>  [1] "week 0"  "week 1"  "week 2"  "week 3"  "week 4"  "week 5"  "week 6" 
#>  [8] "week 7"  "week 8"  "week 9"  "week 10"