Multinomial time-series model.

Takes a list of times, classes and counts, e.g. a COGUK variant like data set with time, (multinomial) class (e.g. variant) and count being the count in that time period. Fits a quadratic B-spline on time to the proportion of the data using nnet::multinom, with approx one degree of freedom per class and per window units of the time series

Usage

multinomial_nnet_model(
  d = i_multinomial_input,
  ...,
  window = 14,
  frequency = "1 day",
  predict = TRUE
)

Arguments

d

A dataframe containing the following columns:

class (factor) - A factor specifying the type of observation. This will be things like

variant, or serotype, for a multinomial model. Any missing data points are ignored.

count (positive_integer) - Positive case counts associated with the specified timeframe

time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a time_period

Must be grouped by: class (exactly).

No default value.

...

not used and present to allow proportion model to be used in a group_modify

window

a number of data points between knots, smaller values result in less smoothing, large value in more.

frequency

the density of the output estimates.

predict

result a prediction. If false we return the model.

Value

a new dataframe with time (as a time period), class, and proportion.0.5, or a model object

Examples

tmp = growthrates::england_covid %>%
  dplyr::filter(date > "2022-01-01") %>%
  growthrates::multinomial_nnet_model(window=21) %>%
  dplyr::glimpse()
#> # weights:  684 (630 variable)
#> initial  value 26851728.826217 
#> iter  10 value 25431321.500081
#> iter  20 value 25353570.627766
#> iter  30 value 25287075.828622
#> iter  40 value 25269813.076045
#> iter  50 value 25265833.484310
#> iter  60 value 25262867.641665
#> iter  70 value 25256092.888323
#> iter  80 value 25244182.901550
#> iter  90 value 25213838.214723
#> iter 100 value 25183443.518406
#> final  value 25183443.518406 
#> stopped after 100 iterations
#> Rows: 13,433
#> Columns: 3
#> Groups: class [19]
#> $ time           <time_prd> 703, 703, 703, 703, 703, 703, 703, 703, 703, 703, …
#> $ class          <fct> 00_04, 05_09, 10_14, 15_19, 20_24, 25_29, 30_34, 35_39,…
#> $ proportion.0.5 <dbl> 0.023767683, 0.038815833, 0.048173657, 0.072622580, 0.1…