Summarise data from a line list to a time-series of counts.

Summarises by group, class and time/date.

time_summarise(
  df,
  unit,
  anchor = "start",
  rectangular = FALSE,
  ...,
  .fill = list(count = 0)
)

Arguments

df

a line list of data you want to summarise, optionally grouped. If this is grouped then each group is treated independently. The remaining columns must contain a date column and may contain a class column.

unit

a period e.g. "1 week"

anchor

one of a date, "start" or "end" or a weekday name e.g. "mon" this will always be one of the start of the time periods we are cutting into

rectangular

should the resulting time series be the same length for all groups. This is only the case if you can be sure that all data has been provided from all otherwise missing data will be treated as zero counts.

...

a spec for a dplyr::summary(...) - optional, and if not provided a count = dplyr::n() or a count = sum(count) is performed.

.fill

a list similar to tidyr::complete for values to fill variables with

Value

a timeseries dataframe grouped by time period,